Parser: store variables in an unordered_map to speed up lookup
Every time we parse a variable, we used to look at a map of vectors
of variables in the current scope. Traversing the whole vector
is very slow, especially if we have lots of variables.

Instead of storing a vector of scope variables in a (ordered) map
(one key per scope), we now store variables in a single flat (unordered) map
with the key of the fully qualified name of the variable (one key per
qualified variable).